legacy/requirements.txt
mutagen
python-slugify
requests-futures